# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UnavailableResponse {
  error: string;
  message: string;
}

export interface GithubRepository {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  htmlUrl: string;
  /** @nullable */
  language: string | null;
  stars: number;
  forks: number;
  updatedAt: string;
  topics: string[];
}

export interface GithubActivity {
  id: string;
  type: string;
  repoName: string;
  repoUrl: string;
  createdAt: string;
  summary: string;
}

export interface GithubContributionDay {
  date: string;
  count: number;
  level: number;
}

export interface GithubContributions {
  total: number;
  days: GithubContributionDay[];
}

export interface LeetcodeProfile {
  username: string;
  /** @nullable */
  totalSolved: number | null;
  /** @nullable */
  easySolved: number | null;
  /** @nullable */
  mediumSolved: number | null;
  /** @nullable */
  hardSolved: number | null;
  /** @nullable */
  contestRating: number | null;
  /** @nullable */
  ranking: number | null;
  /** @nullable */
  acceptanceRate: number | null;
}

export interface LeetcodeProblem {
  id: string;
  /** @nullable */
  number: number | null;
  title: string;
  /** @nullable */
  difficulty: string | null;
  topics: string[];
  /** @nullable */
  solvedAt: string | null;
  url: string;
  /** @nullable */
  solutionUrl: string | null;
}