# lib/api-zod/src/generated/types/githubActivity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GithubActivity {
  id: string;
  type: string;
  repoName: string;
  repoUrl: string;
  createdAt: string;
  summary: string;
}