# lib/api-zod/src/generated/types/githubContributions.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GithubContributionDay } from './githubContributionDay';

export interface GithubContributions {
  total: number;
  days: GithubContributionDay[];
}