# lib/api-zod/src/generated/types/githubRepository.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GithubRepository {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  htmlUrl: string;
  /** @nullable */
  language: string | null;
  stars: number;
  forks: number;
  updatedAt: string;
  topics: string[];
}