# lib/api-zod/src/generated/types/leetcodeProblem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeetcodeProblem {
  id: string;
  /** @nullable */
  number: number | null;
  title: string;
  /** @nullable */
  difficulty: string | null;
  topics: string[];
  /** @nullable */
  solvedAt: string | null;
  url: string;
  /** @nullable */
  solutionUrl: string | null;
}