# lib/api-zod/src/generated/types/leetcodeProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeetcodeProfile {
  username: string;
  /** @nullable */
  totalSolved: number | null;
  /** @nullable */
  easySolved: number | null;
  /** @nullable */
  mediumSolved: number | null;
  /** @nullable */
  hardSolved: number | null;
  /** @nullable */
  contestRating: number | null;
  /** @nullable */
  ranking: number | null;
  /** @nullable */
  acceptanceRate: number | null;
}